requirements.txt
Django==2.2.2
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
celery==4.2.1
django-redis==4.10.0
redis==2.10.6
psycopg2-binary==2.7.5